+++ packages/integration-sdk-runtime/src/execution/utils/getMaskedFields.ts
@@ -1,6 +1,11 @@
-import { IntegrationInvocationConfig } from '@jupiterone/integration-sdk-core';
+import {
+  IntegrationInstanceConfig,
+  IntegrationInvocationConfig,
+} from '@jupiterone/integration-sdk-core';
 import { keys, pickBy } from 'lodash';
 
-export function getMaskedFields(config: IntegrationInvocationConfig) {
+export function getMaskedFields<
+  TIntegrationConfig extends IntegrationInstanceConfig = IntegrationInstanceConfig
+>(config: IntegrationInvocationConfig<TIntegrationConfig>) {
   return keys(pickBy(config.instanceConfigFields, (val) => val.mask));
 }