+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to
 
 ## Unreleased
 
+- [#470](https://github.com/JupiterOne/sdk/issues/470) Introduce
+  `beforeAddEntity` hook into `IntegrationInvocationConfig`. See the development
+  documentation for more information on its usage.
+
 ## 6.1.0 - 2021-05-19
 
 - Upgrade `@jupiterone/data-model@^0.22.1`